refactor(fixture): rename conftest/ directory to fixtures/
